+++ CHANGELOG.md
@@ -8,10 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Support for sandbox applications ([#2178](https://github.com/lbryio/lbry-desktop/pull/2178))
+- Support for sandbox games and applications ([#2178](https://github.com/lbryio/lbry-desktop/pull/2178))
 - CTA to invite page on first run ([#2221](https://github.com/lbryio/lbry-desktop/pull/2221))
 - Responsive related content list ([#2226](https://github.com/lbryio/lbry-desktop/pull/2226))
 - Autoplay content in list of related files ([#2235](https://github.com/lbryio/lbry-desktop/pull/2235))
+- Support for back/forward mouse navigation on Windows ([#2250](https://github.com/lbryio/lbry-desktop/pull/2250))
 
 ### Changed